backend/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
python-jose[cryptography]>=3.3
pandas>=2.0
openpyxl>=3.1
